Moved allowed content types to JSON file and attached it to the http server instance
