+++ lib/http/assets/contenttypes.json
@@ -1,0 +1,53 @@
+{
+    "pdf": "application/pdf",
+    "htm": "text/html",
+    "html": "text/html",
+    "css": "text/css",
+    "js": "text/javascript",
+    "mjs": "text/javascript",
+    "cjs": "text/javascript",
+    "jpg": "image/jpeg",
+    "jpeg": "image/jpeg",
+    "gif": "image/gif",
+    "png": "image/png",
+    "apng": "image/apng",
+    "svg": "image/svg+xml",
+    "svgz": "image/svg+xml",
+    "tif": "image/tiff",
+    "tiff": "image/tiff",
+    "aac": "audio/aac",
+    "abw": "application/x-abiword",
+    "arc": "application/x-freearc",
+    "avif": "image/avif",
+    "avi": "video/x-msvideo",
+    "azw": "application/vnd.amazon.ebook",
+    "bin": "application/octet-stream",
+    "bmp": "image/bmp",
+    "bz": "application/x-bzip",
+    "bz2": "application/x-bzip2",
+    "cda": "application/x-cdf",
+    "csh": "application/x-csh",
+    "csv": "text/csv",
+    "doc": "application/msword",
+    "docx": "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
+    "eot": "application/vnd.ms-fontobject",
+    "epub": "application/epub+zip",
+    "gz": "application/gzip",
+    "ico": "image/vnd.microsoft.icon",
+    "ics": "text/calendar",
+    "jar": "application/java-archive",
+    "json": "application/json",
+    "jsonld": "application/ld+json",
+    "odp": "application/vnd.oasis.opendocument.presentation",
+    "odt": "application/vnd.oasis.opendocument.text",
+    "ods": "application/vnd.oasis.opendocument.spreadsheet",
+    "php": "application/x-httpd-php",
+    "ppt": "application/vnd.ms-powerpoint",
+    "pptx": "application/vnd.openxmlformats-officedocument.presentationml.presentation",
+    "rtf": "application/rtf",
+    "xhtml": "application/xhtml+xml",
+    "xls": "application/vnd.ms-excel",
+    "xlsx": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+    "xml": "application/xml",
+    "zip": "application/zip"
+}
