+++ lib/http/server.go
@@ -4,6 +4,9 @@ import (
 	"log"
 	"net"
 	"strconv"
+	"path/filepath"
+	"errors"
+	"strings"
 )
 
 type HttpServer struct {
@@ -12,6 +15,7 @@ type HttpServer struct {
 	Socket net.Listener
 	SrvLogger *log.Logger
 	StaticRouter FileRoutes
+	AllowedContentTypes map[string]string
 }
 
 func (srv *HttpServer) Static(Route string, TargetPath string) {
@@ -75,7 +79,7 @@ func (srv *HttpServer) handleClient(ClientConnection net.Conn) {
 			srv.sendResponse(httpRequest, httpResponse, ClientConnection.RemoteAddr().String())
 			return
 		}
-		file, err := GetFile(TargetFilePath)
+		file, err := srv.getFile(TargetFilePath)
 		if err != nil {
 			srv.logError(err.Error())
 			httpResponse.Set(StatusInternalServerError, responseVersion, ERROR_MSG_CONTENT_TYPE, StatusInternalServerError.GetErrorContent())
@@ -97,7 +101,7 @@ func (srv *HttpServer) handleClient(ClientConnection net.Conn) {
 			srv.sendResponse(httpRequest, httpResponse, ClientConnection.RemoteAddr().String())
 			return
 		}
-		file, err := GetFile(TargetFilePath)
+		file, err := srv.getFile(TargetFilePath)
 		if err != nil {
 			srv.logError(err.Error())
 			httpResponse.Set(StatusInternalServerError, responseVersion, ERROR_MSG_CONTENT_TYPE, StatusInternalServerError.GetErrorContent())
@@ -125,4 +129,40 @@ func (srv *HttpServer) sendResponse(httpRequest *HttpRequest, httpResponse *Http
 
 func (srv *HttpServer) logError(errorString string) {
 	srv.SrvLogger.Print(errorString)
+}
+
+func (srv *HttpServer) getContentType(CompleteFilePath string) (string, error) {
+	pathType, err := GetPathType(CompleteFilePath)
+	if err != nil {
+		return "", err
+	}
+	if pathType != FILE_TYPE_PATH {
+		return "", errors.New("path provided does not point to a file")
+	}
+	fileExtension := filepath.Ext(CompleteFilePath)
+	if fileExtension == "" {
+		return "", errors.New("file path provided does not contain a file extension")
+	}
+	fileExtension = strings.ToLower(fileExtension)
+	fileMediaType, exists := srv.AllowedContentTypes[fileExtension]
+	if !exists {
+		fileMediaType = "application/octet-stream"
+	}
+	
+	return fileMediaType, nil
+}
+
+func (srv *HttpServer) getFile(CompleteFilePath string) (*File, error) {
+	var file File
+	contentType, err := srv.getContentType(CompleteFilePath)
+	if err != nil {
+		return nil, err
+	}
+	fileContents, err := GetFileContents(CompleteFilePath)
+	if err != nil {
+		return nil, err
+	}
+	file.Contents = fileContents
+	file.ContentType = contentType
+	return &file, nil
 }
